moved orderTypes to payment constants so that it will be readily available for all the components. updated the imports too
