+++ src/components/ServiceDetails/AboutService/ServiceDemo/Purchase/ExpiredSession/GeneralAccountWallet/CreateWallet.js
@@ -1,6 +1,7 @@
 import React from "react";
-import PaymentPopup, { orderTypes } from "./PaymentPopup";
+import PaymentPopup from "./PaymentPopup";
 import { paymentTitles } from "./";
+import { orderTypes } from "../../../../../../../utility/constants/PaymentConstants";
 
 const CreateWallet = ({ visible, setVisibility }) => {
   return (
